LLaVA_GPT4o/requirements.txt
torch
transformers
opencv-python
openai
Pillow
ffmpeg-python
transformers accelerate
bitsandbytes
huggingface_hub
timm==0.9.2
tiktoken>=0.6.0